split sequencer code into separate files
