+++ apps/main-landing/src/components/image-sequencer/image-sequencer.tsx
@@ -1,281 +1,9 @@
 'use client';
 
-import React, { useEffect, useRef, useCallback, useState } from 'react';
-
-type SequencerProps = {
-  images: string[];
-  placeholderImage: string;
-  width?: number;
-  height?: number;
-  fps?: number;
-  direction?: 'forward' | 'backward';
-  infinite?: boolean;
-  startIndex?: number;
-  endIndex?: number;
-  className?: string;
-  onLoad?: () => void;
-  onError?: (error: Error) => void;
-};
-
-const VERTEX_SHADER_SOURCE = `
-  attribute vec2 a_position;
-  attribute vec2 a_texCoord;
-  varying vec2 v_texCoord;
-  void main() {
-    gl_Position = vec4(a_position, 0, 1);
-    v_texCoord = a_texCoord;
-  }
-`;
-
-const FRAGMENT_SHADER_SOURCE = `
-  precision mediump float;
-  uniform sampler2D u_image;
-  varying vec2 v_texCoord;
-  void main() {
-    gl_FragColor = texture2D(u_image, v_texCoord);
-  }
-`;
-
-function createShader(
-  gl: WebGLRenderingContext | WebGL2RenderingContext,
-  type: number,
-  source: string,
-): WebGLShader {
-  const shader = gl.createShader(type);
-  if (!shader) throw new Error('Unable to create shader');
-  gl.shaderSource(shader, source);
-  gl.compileShader(shader);
-  if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-    const info = gl.getShaderInfoLog(shader);
-    gl.deleteShader(shader);
-    throw new Error('Could not compile shader: ' + info);
-  }
-  return shader;
-}
-
-function createProgram(
-  gl: WebGLRenderingContext | WebGL2RenderingContext,
-  vertexShader: WebGLShader,
-  fragmentShader: WebGLShader,
-): WebGLProgram {
-  const program = gl.createProgram();
-  if (!program) throw new Error('Unable to create program');
-  gl.attachShader(program, vertexShader);
-  gl.attachShader(program, fragmentShader);
-  gl.linkProgram(program);
-  if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
-    const info = gl.getProgramInfoLog(program);
-    gl.deleteProgram(program);
-    throw new Error('Could not link program: ' + info);
-  }
-  return program;
-}
-
-// Robust WebGL initialization with fallback: try WebGL2 then fall back to WebGL1.
-function initWebGL(canvas: HTMLCanvasElement) {
-  let gl: WebGLRenderingContext | WebGL2RenderingContext | null = null;
-  try {
-    gl = canvas.getContext('webgl2', {
-      antialias: false,
-      preserveDrawingBuffer: false,
-      powerPreference: 'high-performance',
-    });
-  } catch (e) {
-    // ignore errors here
-  }
-  if (!gl) {
-    gl = canvas.getContext('webgl', {
-      antialias: false,
-      preserveDrawingBuffer: false,
-    });
-  }
-  if (!gl) throw new Error('WebGL is not supported in this browser.');
-
-  const vertexShader = createShader(gl, gl.VERTEX_SHADER, VERTEX_SHADER_SOURCE);
-  const fragmentShader = createShader(gl, gl.FRAGMENT_SHADER, FRAGMENT_SHADER_SOURCE);
-  const program = createProgram(gl, vertexShader, fragmentShader);
-
-  gl.useProgram(program);
-
-  // Create position and texture coordinate buffers
-  const posBuffer = gl.createBuffer();
-  if (!posBuffer) throw new Error('Unable to create position buffer');
-  gl.bindBuffer(gl.ARRAY_BUFFER, posBuffer);
-  gl.bufferData(
-    gl.ARRAY_BUFFER,
-    new Float32Array([-1, -1, 1, -1, -1, 1, 1, 1]),
-    gl.DYNAMIC_DRAW,
-  );
-
-  const texBuffer = gl.createBuffer();
-  if (!texBuffer) throw new Error('Unable to create texture buffer');
-  gl.bindBuffer(gl.ARRAY_BUFFER, texBuffer);
-  gl.bufferData(
-    gl.ARRAY_BUFFER,
-    new Float32Array([0, 1, 1, 1, 0, 0, 1, 0]),
-    gl.STATIC_DRAW,
-  );
-
-  const posLoc = gl.getAttribLocation(program, 'a_position');
-  gl.bindBuffer(gl.ARRAY_BUFFER, posBuffer);
-  gl.enableVertexAttribArray(posLoc);
-  gl.vertexAttribPointer(posLoc, 2, gl.FLOAT, false, 0, 0);
-
-  const texLoc = gl.getAttribLocation(program, 'a_texCoord');
-  gl.bindBuffer(gl.ARRAY_BUFFER, texBuffer);
-  gl.enableVertexAttribArray(texLoc);
-  gl.vertexAttribPointer(texLoc, 2, gl.FLOAT, false, 0, 0);
-
-  return { gl, program, posBuffer };
-}
-
-type TextureData = { texture: WebGLTexture; width: number; height: number };
-
-// Utility to load an image robustly. Uses createImageBitmap when available,
-// otherwise falls back to loading an HTMLImageElement.
-async function loadImage(url: string): Promise<ImageBitmap | HTMLImageElement> {
-  if ('createImageBitmap' in window) {
-    const response = await fetch(url);
-    const blob = await response.blob();
-    return await createImageBitmap(blob);
-  } else {
-    return new Promise((resolve, reject) => {
-      const img = new Image();
-      img.crossOrigin = 'anonymous';
-      img.onload = () => resolve(img);
-      img.onerror = () => reject(new Error('Failed to load image: ' + url));
-      img.src = url;
-    });
-  }
-}
-
-const useTextureLoader = (
-  gl: WebGLRenderingContext | WebGL2RenderingContext | null,
-  images: string[],
-  placeholderImage: string,
-  onError?: (error: Error) => void,
-) => {
-  const [textures, setTextures] = useState<Map<number, TextureData>>(new Map());
-  const [placeholderTexture, setPlaceholderTexture] = useState<TextureData | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-
-  // Load placeholder image first.
-  useEffect(() => {
-    if (!gl) return;
-    const loadPlaceholder = async () => {
-      try {
-        const imageSource = await loadImage(placeholderImage);
-        const texture = gl.createTexture();
-        if (!texture) throw new Error('Unable to create texture');
-
-        gl.bindTexture(gl.TEXTURE_2D, texture);
-        gl.texImage2D(
-          gl.TEXTURE_2D,
-          0,
-          gl.RGBA,
-          gl.RGBA,
-          gl.UNSIGNED_BYTE,
-          imageSource,
-        );
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
-
-        let width, height;
-        if (imageSource instanceof ImageBitmap) {
-          width = imageSource.width;
-          height = imageSource.height;
-          imageSource.close && imageSource.close();
-        } else {
-          width = imageSource.naturalWidth;
-          height = imageSource.naturalHeight;
-        }
-        setPlaceholderTexture({ texture, width, height });
-      } catch (error) {
-        onError?.(error as Error);
-      }
-    };
-    loadPlaceholder();
-  }, [gl, placeholderImage, onError]);
-
-  // Load all other images.
-  useEffect(() => {
-    if (!gl || !placeholderTexture) return;
-    let cancelled = false;
-    const textureMap = new Map<number, TextureData>();
-
-    const loadImageTexture = async (i: number) => {
-      try {
-        if (!images[i]) return;
-        const imageSource = await loadImage(images[i]);
-        const texture = gl.createTexture();
-        if (!texture) throw new Error('Unable to create texture');
-
-        gl.bindTexture(gl.TEXTURE_2D, texture);
-        gl.texImage2D(
-          gl.TEXTURE_2D,
-          0,
-          gl.RGBA,
-          gl.RGBA,
-          gl.UNSIGNED_BYTE,
-          imageSource,
-        );
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
-        gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
-
-        let width, height;
-        if (imageSource instanceof ImageBitmap) {
-          width = imageSource.width;
-          height = imageSource.height;
-          imageSource.close && imageSource.close();
-        } else {
-          width = imageSource.naturalWidth;
-          height = imageSource.naturalHeight;
-        }
-        textureMap.set(i, { texture, width, height });
-      } catch (error) {
-        onError?.(error as Error);
-      }
-    };
-
-    const loadAllImages = async () => {
-      const CONCURRENCY_LIMIT = 30;
-      let index = 0;
-
-      const loadNext = async () => {
-        while (index < images.length) {
-          const currentIndex = index++;
-          await loadImageTexture(currentIndex);
-          await new Promise((resolve) => setTimeout(resolve, 0));
-        }
-      };
-
-      const loaders = [];
-      for (let i = 0; i < CONCURRENCY_LIMIT; i++) {
-        loaders.push(loadNext());
-      }
-
-      await Promise.all(loaders);
-      if (!cancelled) {
-        setTextures(new Map(textureMap));
-        setIsLoading(false);
-      }
-    };
-
-    loadAllImages();
-
-    return () => {
-      cancelled = true;
-      textureMap.forEach((data) => gl.deleteTexture(data.texture));
-      textureMap.clear();
-    };
-  }, [gl, images, onError, placeholderTexture]);
-
-  return { textures, placeholderTexture, isLoading };
-};
+import { useCallback, useEffect, useRef, useState } from 'react';
+import { SequencerProps, TextureData } from './types';
+import { initWebGL } from './web-gl-utils';
+import { useTextureLoader } from './use-texture-loader';
 
 export const ImageSequencer = ({
   images,
@@ -343,7 +71,16 @@ export const ImageSequencer = ({
     const top = 1 - (offsetY / ch) * 2;
     const bottom = 1 - ((offsetY + drawH) / ch) * 2;
 
-    const vertices = new Float32Array([left, bottom, right, bottom, left, top, right, top]);
+    const vertices = new Float32Array([
+      left,
+      bottom,
+      right,
+      bottom,
+      left,
+      top,
+      right,
+      top,
+    ]);
     gl.bindBuffer(gl.ARRAY_BUFFER, posBuffer);
     gl.bufferSubData(gl.ARRAY_BUFFER, 0, vertices);
   }, []);
@@ -370,13 +107,15 @@ export const ImageSequencer = ({
     const current = currentIndexRef.current;
     let next = current;
     if (infinite) {
-      next = direction === 'forward'
-        ? (current + 1) % images.length
-        : (current - 1 + images.length) % images.length;
+      next =
+        direction === 'forward'
+          ? (current + 1) % images.length
+          : (current - 1 + images.length) % images.length;
     } else {
-      next = direction === 'forward'
-        ? Math.min(current + 1, endIndex)
-        : Math.max(current - 1, startIndex);
+      next =
+        direction === 'forward'
+          ? Math.min(current + 1, endIndex)
+          : Math.max(current - 1, startIndex);
     }
     currentIndexRef.current = next;
   }, [direction, images.length, infinite, startIndex, endIndex, isLoading]);
@@ -397,7 +136,13 @@ export const ImageSequencer = ({
 
     animationFrameRef.current = requestAnimationFrame(animate);
     return () => cancelAnimationFrame(animationFrameRef.current);
-  }, [glInitialized, placeholderTexture, frameInterval, updateIndex, drawFrame]);
+  }, [
+    glInitialized,
+    placeholderTexture,
+    frameInterval,
+    updateIndex,
+    drawFrame,
+  ]);
 
   useEffect(() => {
     const gl = glRef.current?.gl;
