+++ apps/main-landing/src/components/image-sequencer/shaders.ts
@@ -1,0 +1,18 @@
+export const VERTEX_SHADER_SOURCE = `
+  attribute vec2 a_position;
+  attribute vec2 a_texCoord;
+  varying vec2 v_texCoord;
+  void main() {
+    gl_Position = vec4(a_position, 0, 1);
+    v_texCoord = a_texCoord;
+  }
+`;
+
+export const FRAGMENT_SHADER_SOURCE = `
+  precision mediump float;
+  uniform sampler2D u_image;
+  varying vec2 v_texCoord;
+  void main() {
+    gl_FragColor = texture2D(u_image, v_texCoord);
+  }
+`;
